Extract out image and pyplot
I think we're done!
